+++ ts/test/constants.ts
@@ -1,0 +1,4 @@
+import { DhtArc } from "@holochain/client";
+
+export const FULL_ARC: DhtArc = [0, 4_294_967_295];
+export const EMPTY_ARC: DhtArc = null;